requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.2
matplotlib>=3.3.4
scikit-learn>=0.24.1
PyYAML>=5.4.1
tqdm>=4.59.0
seaborn>=0.11.1
opencv-python>=4.5.0
pillow>=8.0.0 